+++ src/main/java/pl/telvarost/mojangfixstationapi/client/MojangFixStationApiClientMod.java
@@ -22,6 +22,7 @@ import org.lwjgl.input.Keyboard;
 public class MojangFixStationApiClientMod implements ClientModInitializer {
     public final static KeyBinding COMMAND_KEYBIND = new KeyBinding("Command", Keyboard.KEY_SLASH);
     public final static KeyBinding DEBUG_GRAPH_KEYBIND = new KeyBinding("Debug Graph", Keyboard.KEY_LCONTROL);
+    public static boolean cancelSetScreenNull = false;
 
     @Override
     public void onInitializeClient() {